Added basepeakcapacity and some more RT metrics
Only iRT metrics left from the original SwaMe. I added things like the densities per RT and the TIC per RT because when I made graphs of the results I wanted to see where the trends came from. For example the peakprecision seems to follow the same trends as the TIC.