Empty Arrays can now be read as arrays with zero's
I am not sure if this is the best option to deal with the issue. In this case the empty array (m/z and intensity) is replaced with an array of 5x 0 values. This may throw out other analyses such as density average etc., but a metric was added to notify the user of the amount of missing scans. So I guess the question is whether we should allow the empty array to influence the metrics where an average is calculated.