PR38 Comments 6+7
SVParser used to refer to the different columns by their column number, however, now it looks for the column that contains either the column name in the biognosys irt sheet available on their website, or the specifications in the OpenMS tutorial section 6.4 2016. The idea is still to provide a template that can be filled in in excel for users who would prefer to input the info manually.