Extra run through parser to collect basepeaks added
MzmlReader now loops through the file twice. First to gather info for the basepeaks then to add all the remaining information. With the knowledge of which basepeaks are present, proper chromatogram data can be added.
Also irtbool was added to allow the running of Yamato without irt if necessary.

Chromatogram metrics are not yet calculating correctly. There is also part of the program at the end that appears to be running twice. More investigation is needed.