Small changes to make chromatogrammetrics work
CrawDad now takes as input only the parts of the code  that are relevant to each instance of the basepeak.bpRT, not every time the basepeak is picked up. So instead of having one large chromatogram for the entire RT, we have smaller chromatograms related to each time said basepeak was picked up as a basepeak of a scan.

New class written within Run - analysissettings that contains masstolerance and rtTolerance.

Also protected against instances where the scan is empty in the basepeakparser.