CombineFilesAtTheEnd
When the last file is finished, SwaMe now combines all the tsv's of the same group it has output into one combined file. E.g. if two inputfiles a and b were run, there would be iRTMetrics_a.tsv and iRTMetrics_b.tsv in the same folder. Now there will also be a file AllIRTMetrics.tsv which is a combination of the two.