PredictedChargeStateOneProportionAdded
Working from the premise that M and M+1 pairs are singly charged and the assumption that all peaks that are 1 +- massTolerance apart are M and M+1 pairs; we calculate the difference between each ion in the mz array and each previous ion. 
An array is created of cumulative sums of those differences and each new difference is added to the cumulate of all of the ions before it, until the ion reaches/overshoots 1+-massTolerance. If the cumulative difference is within our range, both the ion to which the cusum belongs and the current ion in the loop is added to a list of matched ions(indexes) and we also stop adding to that ion in future. At the end all the distinct items in indexes are counted and that count is divided by the total amount of floats in the original mzs array to get the proportion.