CycleTimes calculation more precise and in seconds
Previously cycle times was generalized by dividing the number of cycles run by the time it took, now the exact cycle time is calculated for each cycle and out of that array, the average is taken. The value is a little off as the end of the cycle is considered the start of the last scan of the cycle, so the last scan is not included. This approach was taken to be consequent with regards to the last cycle of each segment.

This value is also multiplied by 60 to obtain the answer in seconds rather than minutes.