Changes to deal with non-sequential windows
I found a file where the isolation windows were fixed and the target mzs were in the order 400.83...810.23 then 410.83 ... 820.23, then back to 400.83...810.23. By sorting the isolation windows by the scanstarttime before dividing into swaths and then changing the number of swaths to the total distinct isolation windows rather than the swathnumber of the swath with the largest target ion, I was able to keep the configuration and not order the swaths 400.83, 410.83 etc. as well as including all swaths even if their  isolation target mz is smaller than the previous swath.