ChromatogramMetricGeneratorTestAddedToCheckIntensities
The body of chromatogrammetricGenerator was also altered slightly as all the different functions used the  intensities and starttimes arrays this property was added to the class. This allows the property to also be tested in the unit test.

The unit test tests that the intensities array selected the correct values via a lambda expression and that the interpolation function correctly performed the interpolation.