Refactor MzmlParser to remove SwaMe and Prognosticator additions
This adds a minimal plugin architecture to MzmlReader (see class comment) and introduces a new project, SwaMe.Pipeline, as the plug-in for SwaMe and Prognosticator.  It also makes Scan and Run extensible by parameterising the MzmlReader and using a factory pattern for concrete scan and run generation.

Changing Ms2Scans from ConcurrentBag to List changes the order of enumeration of the MS2 scans.  It appears that the results of at least some metrics depended on the order (they shouldn't - multi-threaded parsing introduces inherent uncertainties into the enumeration order).  I strongly recommend that metric authors re-run their tests on real data and investigate any discrepancies with previous results.