Prevent some simple race conditions when calculating BasePeaks.
This merely prevents errors due to reading an intermediate state (the A in ACID transactions).  Note that a more complex race condition is still present: there is non-determinism in whether a given BasePeak or BpkRt is created based on the state of the run at that time.  This patch does not (and cannot) address that.